Migrate storage to dom
